requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI and LangChain
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.1.0
sentence-transformers>=2.2.0
chromadb>=0.4.18

# Database
sqlalchemy>=2.0.23

# OCR
pdf2image>=1.17.0
pytesseract>=0.3.10

# File Processing
pypdf2>=3.0.1
python-docx>=1.1.0
markdown>=3.5.1

# Data Validation
pydantic>=2.5.0

# Other Tools
python-dotenv==1.0.0
requests==2.31.0
